# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pabawi - Unified Remote Execution Interface API
  description: |
    REST API for Pabawi, a web-based interface for managing Bolt automation with integrated
    PuppetDB, Puppetserver, and Hiera support. This API provides endpoints for managing
    inventory from multiple sources, executing commands and tasks, gathering facts,
    running Puppet, installing packages, viewing execution history, and analyzing Hiera data.

    ## Integration Sources

    The API supports multiple integration sources:
    - **Bolt**: Direct execution and inventory management
    - **PuppetDB**: Node inventory, facts, reports, catalogs, and events
    - **Puppetserver**: CA inventory, facts, catalog compilation, and environments
    - **Hiera**: Configuration data lookup and analysis

    ## Expert Mode

    Many endpoints support an "expert mode" that provides additional diagnostic information
    when errors occur. To enable expert mode, include the `X-Expert-Mode: true` header in your
    request. When enabled, error responses will include:
    - Full stack traces
    - Request IDs for correlation
    - Execution context
    - Raw Bolt CLI output

    ## Streaming Execution Output

    For long-running operations, you can subscribe to real-time execution output via
    Server-Sent Events (SSE). After starting an execution, connect to the streaming
    endpoint to receive stdout, stderr, and status updates as they occur.

  version: 0.4.0
  contact:
...
  license:
...
servers:
...
paths:
...
  /nodes/{id}/install-package:
    post:
      summary: Install package on node
      description: |
        Install a package on a target node using a configured package installation task.

        Returns immediately with an execution ID. Use the executions endpoint to
        retrieve results, or subscribe to the streaming endpoint for real-time output.
      tags:
        - Packages
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - $ref: '#/components/parameters/ExpertModeHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - taskName
                - packageName
              properties:
                taskName:
                  type: string
                  description: Name of the package installation task
                  example: tp::install
                packageName:
                  type: string
                  description: Name of the package to install
                  example: nginx
                ensure:
                  type: string
                  enum: [present, absent, latest]
                  description: Desired package state
                  default: present
                version:
                  type: string
                  description: Specific package version (optional)
                  example: 1.18.0
                settings:
                  type: object
                  description: Additional package settings
                  additionalProperties: true
                expertMode:
                  type: boolean
                  description: Enable expert mode for detailed output
                  default: false
      responses:
        '202':
          description: Package installation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionStarted'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Node not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWithExpertMode'
...
components:
  parameters:
    NodeId:
      name: id
      in: path
      required: true
      description: Node identifier (name or ID)
      schema:
        type: string
        example: node1
    ExpertModeHeader:
      name: X-Expert-Mode
      in: header
      description: Enable expert mode for detailed error output
      schema:
        type: boolean
        default: false
  schemas:
...
    ExecutionStarted:
      type: object
      properties:
        executionId:
          type: string
          description: Unique execution identifier
        status:
          type: string
          enum: [running]
        message:
          type: string
      example:
        executionId: abc123def456  # pragma: allowlist secret
        status: running
        message: Command execution started
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              description: Additional error details
      example:
        error:
          code: INVALID_REQUEST
          message: Request validation failed
          details: Command is required
    ErrorWithExpertMode:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              description: Additional error details
            stackTrace:
              type: string
              description: Full stack trace (expert mode only)
            requestId:
              type: string
              description: Request correlation ID (expert mode only)
            timestamp:
              type: string
              format: date-time
              description: Error timestamp (expert mode only)
            rawResponse:
              description: Raw Bolt CLI output (expert mode only)
            executionContext:
              type: object
              description: Execution context (expert mode only)
              properties:
                endpoint:
                  type: string
                method:
                  type: string
                requestId:
                  type: string
      example:
        error:
          code: BOLT_EXECUTION_FAILED
          message: Bolt command failed
          details: Connection timeout
          stackTrace: "Error: Bolt command failed\n    at BoltService.runCommand..."
          requestId: req-abc123
          timestamp: "2024-01-01T00:00:00.000Z"
          rawResponse: "Error: Connection timeout after 30s"
          executionContext:
            endpoint: /api/nodes/node1/command
            method: POST
            requestId: req-abc123
...
  securitySchemes:
    # Placeholder for future authentication
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication (not implemented in v0.1.0)

# Security is not enforced in v0.1.0
security: []